Crate mirror1

This documentation was generated from mirror crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

accounts (insert)

contacts (delete, get, insert, list, patch and update)

locations (get and list)

settings (get)

subscriptions (delete, insert, list and update)

timeline (attachments.delete, attachments.get, attachments.insert, attachments.list, delete, get, insert, list, patch and update)

Everything else about the mirror v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.contacts().insert(...).doit()
let r = hub.contacts().delete(...).doit()
let r = hub.contacts().list(...).doit()
let r = hub.contacts().update(...).doit()
let r = hub.contacts().patch(...).doit()
let r = hub.contacts().get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
mirror1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate mirror1;
use mirror1::Contact;
use mirror1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mirror1::Mirror;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Mirror::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Contact = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.contacts().update(&req, "id")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Account
Represents an account passed into the Account Manager on Glass.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert accounts (request|response)

AccountInsertMethodBuilder
Inserts a new account for a user

AccountMethodsBuilder
A builder providing access to all methods supported on account resources. It is not used directly, but through the Mirror hub.

Attachment
Represents media content, such as a photo, that can be attached to a timeline item.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

attachments.insert timeline (response)

attachments.get timeline (response)

AttachmentsListResponse
A list of Attachments. This is the response from the server to GET requests on the attachments collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

attachments.list timeline (response)

AuthToken
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Command
A single menu command that is part of a Contact.

This type is not used in any activity, and only used as part of another schema.

Contact
A person or group that can be used as a creator or a contact.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert contacts (request|response)

delete contacts (none)

list contacts (none)

update contacts (request|response)

patch contacts (request|response)

get contacts (response)

ContactDeleteMethodBuilder
Deletes a contact.

ContactGetMethodBuilder
Gets a single contact by ID.

ContactInsertMethodBuilder
Inserts a new contact.

ContactListMethodBuilder
Retrieves a list of contacts for the authenticated user.

ContactMethodsBuilder
A builder providing access to all methods supported on contact resources. It is not used directly, but through the Mirror hub.

ContactPatchMethodBuilder
Updates a contact in place. This method supports patch semantics.

ContactUpdateMethodBuilder
Updates a contact in place.

ContactsListResponse
A list of Contacts representing contacts. This is the response from the server to GET requests on the contacts collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list contacts (response)

Location
A geographic location that can be associated with a timeline item.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get locations (response)

list locations (none)

LocationGetMethodBuilder
Gets a single location by ID.

LocationListMethodBuilder
Retrieves a list of locations for the user.

LocationMethodsBuilder
A builder providing access to all methods supported on location resources. It is not used directly, but through the Mirror hub.

LocationsListResponse
A list of Locations. This is the response from the server to GET requests on the locations collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list locations (response)

MenuItem
A custom menu item that can be presented to the user by a timeline item.

This type is not used in any activity, and only used as part of another schema.

MenuValue
A single value that is part of a MenuItem.

This type is not used in any activity, and only used as part of another schema.

Mirror
Central instance to access all Mirror related resource activities

Notification
A notification delivered by the API.

This type is not used in any activity, and only used as part of another schema.

NotificationConfig
Controls how notifications for a timeline item are presented to the user.

This type is not used in any activity, and only used as part of another schema.

Setting
A setting for Glass.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get settings (response)

SettingGetMethodBuilder
Gets a single setting by ID.

SettingMethodsBuilder
A builder providing access to all methods supported on setting resources. It is not used directly, but through the Mirror hub.

Subscription
A subscription to events on a collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert subscriptions (request|response)

delete subscriptions (none)

list subscriptions (none)

update subscriptions (request|response)

SubscriptionDeleteMethodBuilder
Deletes a subscription.

SubscriptionInsertMethodBuilder
Creates a new subscription.

SubscriptionListMethodBuilder
Retrieves a list of subscriptions for the authenticated user and service.

SubscriptionMethodsBuilder
A builder providing access to all methods supported on subscription resources. It is not used directly, but through the Mirror hub.

SubscriptionUpdateMethodBuilder
Updates an existing subscription in place.

SubscriptionsListResponse
A list of Subscriptions. This is the response from the server to GET requests on the subscription collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list subscriptions (response)

TimelineAttachmentDeleteMethodBuilder
Deletes an attachment from a timeline item.

TimelineAttachmentGetMethodBuilder
Retrieves an attachment on a timeline item by item ID and attachment ID.

TimelineAttachmentInsertMethodBuilder
Adds a new attachment to a timeline item.

TimelineAttachmentListMethodBuilder
Returns a list of attachments for a timeline item.

TimelineDeleteMethodBuilder
Deletes a timeline item.

TimelineGetMethodBuilder
Gets a single timeline item by ID.

TimelineInsertMethodBuilder
Inserts a new item into the timeline.

TimelineItem
Each item in the user's timeline is represented as a TimelineItem JSON structure, described below.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

patch timeline (request|response)

insert timeline (request|response)

get timeline (response)

update timeline (request|response)

TimelineListMethodBuilder
Retrieves a list of timeline items for the authenticated user.

TimelineListResponse
A list of timeline items. This is the response from the server to GET requests on the timeline collection.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list timeline (response)

TimelineMethodsBuilder
A builder providing access to all methods supported on timeline resources. It is not used directly, but through the Mirror hub.

TimelinePatchMethodBuilder
Updates a timeline item in place. This method supports patch semantics.

TimelineUpdateMethodBuilder
Updates a timeline item in place.

UserAction
Represents an action taken by the user that triggered a notification.

This type is not used in any activity, and only used as part of another schema.

UserData
There is no detailed description.

This type is not used in any activity, and only used as part of another schema.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.