Struct mirror1::TimelineMethodsBuilder

pub struct TimelineMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on timeline resources. It is not used directly, but through the Mirror hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate mirror1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use mirror1::Mirror;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Mirror::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `attachments_delete(...)`, `attachments_get(...)`, `attachments_insert(...)`, `attachments_list(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.timeline();

Methods

impl<'a, C, NC, A> TimelineMethodsBuilder<'a, C, NC, A>

fn attachments_list(&self, item_id: &str) -> TimelineAttachmentListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a list of attachments for a timeline item.

fn insert(&self, request: &TimelineItem) -> TimelineInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Inserts a new item into the timeline.

fn patch(&self, request: &TimelineItem, id: &str) -> TimelinePatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a timeline item in place. This method supports patch semantics.

fn list(&self) -> TimelineListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of timeline items for the authenticated user.

fn attachments_insert(&self, item_id: &str) -> TimelineAttachmentInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds a new attachment to a timeline item.

fn attachments_delete(&self, item_id: &str, attachment_id: &str) -> TimelineAttachmentDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes an attachment from a timeline item.

fn delete(&self, id: &str) -> TimelineDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a timeline item.

fn update(&self, request: &TimelineItem, id: &str) -> TimelineUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a timeline item in place.

fn attachments_get(&self, item_id: &str, attachment_id: &str) -> TimelineAttachmentGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves an attachment on a timeline item by item ID and attachment ID.

fn get(&self, id: &str) -> TimelineGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Gets a single timeline item by ID.